+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+1.1 (2013-??-??)
+================
+ - Added `atmos.utils.Logger`, a super-thin wrapper around `java.util.logging`.
+ - Added `atmos.utils.Encoding`, an interface for converting arbitrary types to and from strings.
+ - Added `atmos.utils.Attributes`, a template class for defining type-safe key/value pairs based on maps of strings.
+
 1.0 (2013-11-21)
 ================
-
 Initial release.